LF:  move the type ordering form interpreter to language package
CHANGELOG_BEGIN
CHANGELOG_END
